requirements.txt
asyncpraw==7.7.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
praw==7.7.1
anthropic==0.18.1
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0 